+++ src/loader/interfaces.c
@@ -2,8 +2,8 @@
 #include<malloc.h>
 #include<interfaces.h>
 
-uint16_t parseInterfaces(FILE *r, uint16_t interfaceCount){
-    uint16_t interface_ptr = heapAlloc(sizeof(uint16_t) * interfaceCount);
+int parseInterfaces(FILE *r, uint16_t interfaceCount){
+    int interface_ptr = heapAlloc(sizeof(uint16_t) * interfaceCount);
     uint16_t *interface = (uint16_t*)&heap[interface_ptr];
 
     for (int i = 0; i<interfaceCount; i++){
